wip: update selected school / university page
